[refactor] Move types into types folder, reorganize accordingly
