+++ client.go
@@ -1,8 +1,9 @@
-package aptos
+package aptos_go_sdk
 
 import (
 	"errors"
 	"github.com/aptos-labs/aptos-go-sdk/core"
+	"github.com/aptos-labs/aptos-go-sdk/types"
 	"time"
 )
 
@@ -112,12 +113,12 @@ func (client *Client) SetTimeout(timeout time.Duration) {
 }
 
 // Info Retrieves the node info about the network and it's current state
-func (client *Client) Info() (info NodeInfo, err error) {
+func (client *Client) Info() (info types.NodeInfo, err error) {
 	return client.nodeClient.Info()
 }
 
 // Account Retrieves information about the account such as SequenceNumber and AuthenticationKey
-func (client *Client) Account(address core.AccountAddress, ledgerVersion ...int) (info AccountInfo, err error) {
+func (client *Client) Account(address core.AccountAddress, ledgerVersion ...int) (info types.AccountInfo, err error) {
 	return client.nodeClient.Account(address, ledgerVersion...)
 }
 
@@ -129,12 +130,12 @@ func (client *Client) AccountResource(address core.AccountAddress, resourceType 
 
 // AccountResources fetches resources for an account into a JSON-like map[string]any in AccountResourceInfo.Data
 // For fetching raw Move structs as BCS, See #AccountResourcesBCS
-func (client *Client) AccountResources(address core.AccountAddress, ledgerVersion ...int) (resources []AccountResourceInfo, err error) {
+func (client *Client) AccountResources(address core.AccountAddress, ledgerVersion ...int) (resources []types.AccountResourceInfo, err error) {
 	return client.nodeClient.AccountResources(address, ledgerVersion...)
 }
 
 // AccountResourcesBCS fetches account resources as raw Move struct BCS blobs in AccountResourceRecord.Data []byte
-func (client *Client) AccountResourcesBCS(address core.AccountAddress, ledgerVersion ...int) (resources []AccountResourceRecord, err error) {
+func (client *Client) AccountResourcesBCS(address core.AccountAddress, ledgerVersion ...int) (resources []types.AccountResourceRecord, err error) {
 	return client.nodeClient.AccountResourcesBCS(address, ledgerVersion...)
 }
 
@@ -182,7 +183,7 @@ func (client *Client) Transactions(start *uint64, limit *uint64) (data []map[str
 }
 
 // SubmitTransaction Submits an already signed transaction to the blockchain
-func (client *Client) SubmitTransaction(signedTransaction *SignedTransaction) (data map[string]any, err error) {
+func (client *Client) SubmitTransaction(signedTransaction *types.SignedTransaction) (data map[string]any, err error) {
 	return client.nodeClient.SubmitTransaction(signedTransaction)
 }
 
@@ -199,13 +200,13 @@ func (client *Client) Fund(address core.AccountAddress, amount uint64) error {
 
 // BuildTransaction Builds a raw transaction from the payload and fetches any necessary information
 // from onchain
-func (client *Client) BuildTransaction(sender core.AccountAddress, payload TransactionPayload, options ...any) (rawTxn *RawTransaction, err error) {
+func (client *Client) BuildTransaction(sender core.AccountAddress, payload types.TransactionPayload, options ...any) (rawTxn *types.RawTransaction, err error) {
 	return client.nodeClient.BuildTransaction(sender, payload, options...)
 }
 
 // BuildSignAndSubmitTransaction Convenience function to do all three in one
 // for more configuration, please use them separately
-func (client *Client) BuildSignAndSubmitTransaction(sender *core.Account, payload TransactionPayload, options ...any) (hash string, err error) {
+func (client *Client) BuildSignAndSubmitTransaction(sender *core.Account, payload types.TransactionPayload, options ...any) (hash string, err error) {
 	return client.nodeClient.BuildSignAndSubmitTransaction(sender, payload, options...)
 }
