+++ client_util.go
@@ -1,9 +1,10 @@
-package aptos
+package aptos_go_sdk
 
 import (
 	"encoding/binary"
 	"fmt"
 	"github.com/aptos-labs/aptos-go-sdk/core"
+	"github.com/aptos-labs/aptos-go-sdk/types"
 	"net/url"
 	"runtime/debug"
 )
@@ -51,18 +52,18 @@ func init() {
 // Amount in Octas (10^-8 APT)
 //
 // options may be: MaxGasAmount, GasUnitPrice, ExpirationSeconds, ValidUntil, SequenceNumber, ChainIdOption
-func APTTransferTransaction(client *Client, sender *core.Account, dest core.AccountAddress, amount uint64, options ...any) (signedTxn *SignedTransaction, err error) {
+func APTTransferTransaction(client *Client, sender *core.Account, dest core.AccountAddress, amount uint64, options ...any) (signedTxn *types.SignedTransaction, err error) {
 	var amountBytes [8]byte
 	binary.LittleEndian.PutUint64(amountBytes[:], amount)
 
 	rawTxn, err := client.BuildTransaction(sender.Address,
-		TransactionPayload{Payload: &EntryFunction{
-			Module: ModuleId{
+		types.TransactionPayload{Payload: &types.EntryFunction{
+			Module: types.ModuleId{
 				Address: core.AccountOne,
 				Name:    "aptos_account",
 			},
 			Function: "transfer",
-			ArgTypes: []TypeTag{},
+			ArgTypes: []types.TypeTag{},
 			Args: [][]byte{
 				dest[:],
 				amountBytes[:],
