+++ nodeClient_test.go
@@ -1,4 +1,4 @@
-package aptos
+package aptos_go_sdk
 
 import (
 	"testing"
