+++ types/moduleId.go
@@ -1,0 +1,20 @@
+package types
+
+import (
+	"github.com/aptos-labs/aptos-go-sdk/bcs"
+	"github.com/aptos-labs/aptos-go-sdk/core"
+)
+
+type ModuleId struct {
+	Address core.AccountAddress
+	Name    string
+}
+
+func (mod *ModuleId) MarshalBCS(bcs *bcs.Serializer) {
+	mod.Address.MarshalBCS(bcs)
+	bcs.WriteString(mod.Name)
+}
+func (mod *ModuleId) UnmarshalBCS(bcs *bcs.Deserializer) {
+	mod.Address.UnmarshalBCS(bcs)
+	mod.Name = bcs.ReadString()
+}
