+++ types/nodeInfo_test.go
@@ -1,4 +1,4 @@
-package aptos
+package types
 
 import (
 	"bytes"
