+++ types/typetag_test.go
@@ -1,4 +1,4 @@
-package aptos
+package types
 
 import (
 	"github.com/aptos-labs/aptos-go-sdk/core"
